+++ tests/unittests/classification/test_precision_at_fixed_recall.py
@@ -21,7 +21,7 @@ from scipy.special import expit as sigmoid
 from scipy.special import softmax
 from sklearn.metrics import precision_recall_curve as sk_precision_recall_curve
 
-from torchmetrics.classification.precision_at_fixed_recall import (
+from torchmetrics.classification.precision_fixed_recall import (
     BinaryPrecisionAtFixedRecall,
     MulticlassPrecisionAtFixedRecall,
     MultilabelPrecisionAtFixedRecall,
